requirements.txt
# LLM Training Data Pipeline Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Wikipedia Parsing
mwparserfromhell>=0.6.5
bz2file>=0.98

# Text Processing
regex>=2023.10.3
ftfy>=6.1.1
unidecode>=1.3.7

# Deduplication
datasketch>=1.6.4
xxhash>=3.4.1

# Language Detection
langdetect>=1.0.9

# Tokenization
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Quality Filtering
kenlm>=0.2.0  # Optional: for perplexity scoring

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
requests>=2.31.0
click>=8.1.7

# Logging & Metrics
rich>=13.6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
